feat: Move date formatting to utils
