+++ packages/frontend/editor-ui/src/features/execution/insights/components/InsightsDataRangePicker.vue
@@ -8,13 +8,9 @@ import type {
 	N8nDateRangePickerRootEmits,
 } from '@n8n/design-system';
 import { N8nButton, N8nDateRangePicker, N8nIcon } from '@n8n/design-system';
-import dateformat from 'dateformat';
 import { computed, ref, shallowRef, watch } from 'vue';
-import { isPresetRange } from '../insights.utils';
+import { formatDateRange, isPresetRange } from '../insights.utils';
 import InsightsUpgradeModal from './InsightsUpgradeModal.vue';
-
-const DATE_FORMAT_DAY_MONTH_YEAR = 'd mmm, yyyy';
-const DATE_FORMAT_DAY_MONTH = 'd mmm';
 
 type Props = Pick<N8nDateRangePickerProps, 'maxValue' | 'minValue'>;
 type Value = {
@@ -136,18 +132,7 @@ const formattedRange = computed(() => {
 
 	if (!start) return 'Select range';
 
-	const startStr = start.toString();
-	const endStr = end?.toString();
-
-	if (!end || startStr === endStr) {
-		return dateformat(startStr, DATE_FORMAT_DAY_MONTH_YEAR);
-	}
-
-	if (start.year === end.year) {
-		return `${dateformat(startStr, DATE_FORMAT_DAY_MONTH)} - ${dateformat(endStr, DATE_FORMAT_DAY_MONTH_YEAR)}`;
-	}
-
-	return `${dateformat(startStr, DATE_FORMAT_DAY_MONTH_YEAR)} - ${dateformat(endStr, DATE_FORMAT_DAY_MONTH_YEAR)}`;
+	return formatDateRange({ start, end });
 });
 
 const isActiveRange = (presetValue: number) =>
